backend/services/requirements.txt
flask
pandas
numpy
requests
pg8000
scikit-learn
xgboost
gunicorn
python-dotenv
